+++ NEWS.md
@@ -2,4 +2,4 @@
 
 * Development Version
 
-* PK Parameter Calculation
+* Added PK Parameter Calculation